==> AirBnB Price Prediction Challenge/1. Feature Engineering_D.ipynb <==
### Imports

# Importing essential libraries

# pandas and numpy for data manipulation
import numpy as np
import pandas as pd

# matplotlib and seaborn for visuilization
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

# Display all the columns of the dataframe
pd.pandas.set_option('display.max_columns',None)

# No warnings
import warnings
warnings.filterwarnings('ignore')

### Load in the Data and Examine

# %%HTML
# <style type="text/css">
# table.dataframe td, table.dataframe th {
#     border: 1px  black solid !important;
#   color: black !important;
# }
# </style>

# Read in data into dataframe
file_train=r'D:\Deloitte\case study\train.csv'
file_test=r'D:\Deloitte\case study\test.xlsx'

ab_train=pd.read_csv(file_train)
ab_test=pd.read_excel(file_test)

# Returns the number of Rows and Columns in train data
print('shape of train data: {}'.format(ab_train.shape))

# Returns the number of Rows and Columns in test data
print('shape of test data: {}'.format(ab_test.shape))

# ab_test does not have the outcome variable 
# In order to combine the two we will add the response variable to ab_test

ab_test['log_price']=np.nan

* combine the two datasets to pre-process the data together since the data on which the model is trained should
 undergo the same preprocessing as the data on which the predictions are made

# creating identifier column for train and test
ab_train['data']='train'
ab_test['data']='test'

# the columns in the two data frames should be in the same order to enable concatenation
ab_test=ab_test[ab_train.columns]

ab_all=pd.concat([ab_train,ab_test],axis=0)

# Returns top 5 rows from the dataset
ab_all.head()

ab=ab_all.copy()

# Returns the number of Rows and Columns of overall data[both train and test]
print('shape of overall data: {}'.format(ab.shape))

###### 1. 'id' ---> drop this column while model building

###### 2. 'property_type' ----> create dummies based on dependent variable

###### 3. 'room_type' ---> create dummies

###### 4. 'amenities' --->Drop in this notebook. we will create MultiLabelBinarizer in 'amenities' separate notebook.
* we concat both and use them in final Model

###### 5. 'accommodates' --->  let it be

###### 6.  'bathrooms' ---> bathrooms cannot be 1.5,2.5.... 
* so i will consider 0.5 to 1,1.5 to 2,2.5 to 3,3.5 to 4,4.5 to 5,5.5 to 6,6.5 and 7.5 to 7
* Replace missing value to 0.0 bathrooms

###### 7.  'bed_type' ---> create Futon,Pull-out Sofa, Airbed, Couch to 'other' and then create dummies

###### 8. 'cancellation_policy'--->super_strict_30 and super_strict_60 to 'super_strict' and then create dummies

###### 9. 'cleaning_fee' ---> It is boolean data type.
* In feature engineering, we will create True to 1 and False to 0

###### 10. 'city' ---> Due to less frequency count in some variables. lets club some features like
* SF and DC which has >3000 frequency count to one variable ie., SF_DC
* Chicago and Boston which has <3000 frequency count to one variable ie.,chicago_boston
* Then create dummies

###### 11. 'description' ---> lets try NLP by using tfidf or count vectorizer. unstructured data 

###### 12. 'first_review' ---> Both first_review and last_review have almost 21.4% values are Missing values.
* Since they are directly correlated we will drop first_review and make use last review.

###### 13. 'host_has_profile_pic' ---> I think people use different images like celebrity photo and some use their own photo.
* It will be Imbalanced because one category frequency 99.4% of data.
* I think its better to drop
* Beacause once the host identity is verified then only it will allows the host to book

###### 14. 'host_identity_verified' ---> I am assuming 't' tobe True or correct match and 'f' to be False or Incorrect match
* Replace missing value with f 
* lets create t to 1 and f to 0

###### 15. 'host_response_rate' Replace '%' with '' (empty space) and then convert it numeric
* It has almost 25% of the missing values. 
* create indicator column missing-->0, non-missing-->1

###### 16. 'host_since' ---> ' lets covert into date-time format and extract features out of it and 
* Replace missing with min year

###### 17. 'instant_bookable' ---> lets create  t to 1 and f to 0
* In cross table, we can see frequency count of combinations.
* It is giving me like instant_bookable f and host_identity_verified f is 11098 which is almost 22%.
* we have to take care of instant_bookable f and host_identity_verified t is 25671 which is almost 51%. means we have to instant booking availability. so that we can make profits out of it

###### 18. 'last_review' --->Since there are almost 21.4% values are Missing.
* replace missing with min 

###### 19. 'latitude' --> convert to x,y,z co-ordinates
###### 20. 'longitude' ---> convert to x,y,z co-ordinates
* x=cos(lat)*cos(lon)
* y=cos(lat)*sin(lon)
* z=sin(lat)

###### 21. 'name' ---> lets try NLP by using tfidf or count vectorizer 

###### 22. 'neighbourhood'--->Drop this variable since it has lot of unique values

###### 23. 'number of reviews' ---> let it be

###### 24. 'review_scores_rating' --->Since there are almost 22.6% values are Missing.
* Replace  missing-->0

###### 25. 'thumbnail_url' ---> for building Ml model, lets drop this variable
* we can CNN Deep Learning. But for now lets drop this column

###### 26. 'zipcode' --->  Lets drop this variable. 
* Make use of latitude, longitude 

###### 27. 'bedrooms' ---> Replace missing value with 0.0

###### 28. 'beds' ---> Replace missing value with 0.0

###### 29. 'log_price' ---> I am assuming that log_price is log transformed prices. Target variable.

###### 30. ' ' ---> drop this column. It is junk and the whole column has NAN values.

print('shape of data before: {}'.format(len(ab.columns)))
ab=ab.drop([' ','amenities','first_review','host_has_profile_pic', 'zipcode',
             'description', 'name','neighbourhood','thumbnail_url'],axis=1)
print('shape of data after: {}'.format(len(ab.columns)))

ab.columns

# Returns basic information on all columns
ab.info()

ab.isnull().sum()

# property_type ---> create dummies based on some frequency cut off
ab['property_type'].value_counts()

# 'bathrooms' ---> bathrooms cannot be 1.5,2.5.... 
# so i will club or consider 0.5 to 1,1.5 to 2,2.5 to 3,3.5 to 4,4.5 to 5,5.5 to 6,6.5 and 7.5 to 8
# Replace missing value to 0.0 bathrooms

ab['bathrooms']=ab['bathrooms'].replace({0.5:1,1.5:2,2.5:3,3.5:4,4.5:5,5.5:6,6.5:7,7.5:8})
ab['bathrooms']=np.where(ab['bathrooms'].isnull(),0.0,ab['bathrooms'])

# 'bed_type'---> create Futon,Pull-out Sofa, Airbed, Couch to 'other' and then create dummies
ab.loc[~ab['bed_type'].isin(['Real Bed']), 'bed_type'] = 'Other'
ab['bed_type'].value_counts()

# 'cancellation_policy'--->super_strict_30 and super_strict_60 to 'super_strict' and then create dummies
ab.loc[ab['cancellation_policy'].isin(['super_strict_30','super_strict_60']),
         'cancellation_policy']='super_strict'
ab['cancellation_policy'].value_counts()

# 'cleaning_fee' ---> It is boolean data type. we will create True to 1 and False to 0
ab['cleaning_fee']=np.where(ab['cleaning_fee']==True,1,0)
ab['cleaning_fee'].value_counts()

# 'city'---> Due to less freqency count in some variables. lets club some features like
# SF and DC which has >5000 frequency count to one variable ie., SF_DC
# Chicago and Boston which has <5000 frequency count to one variable ie., other
# create dummies
ab.loc[ab['city'].isin(['SF','DC']),'city']='SF_DC'
ab.loc[ab['city'].isin(['Chicago','Boston']),'city']='chicago_boston'
ab['city'].value_counts()

# 'host_identity_verified'---> I am assuming t tobe True or correct match and f to be False or Incorrect match
# Replace missig value with f 
# lets create  t to 1 and f to 0
ab['host_identity_verified']=np.where(ab['host_identity_verified'].isnull(),'f',ab['host_identity_verified'])
ab['host_identity_verified']=np.where(ab['host_identity_verified']=='t',1,0)
ab['host_identity_verified'].value_counts()

# 'host_response_rate'---> Replace '%' with '' (empty space) and then convert it numeric
# It has almost 25% of the missing values. 
# create indicator column missing-->0, non-missing-->1
ab['host_response_rate']=pd.to_numeric(ab['host_response_rate'].str.replace('%',''))
ab['host_response_rate']=np.where(ab['host_response_rate'].isnull(),0,1)
ab['host_response_rate'].value_counts()

# 'host_since'--> lets covert into date-time format and extract features out of it and 
# Replace missing with min
ab['host_since']=pd.to_datetime(ab['host_since'],format='%m/%d/%Y',errors='coerce').dt.year
ab.loc[ab['host_since'].isnull(),'host_since']=ab.loc[ab['data']=='train','host_since'].min()
ab['host_since'].value_counts()

# 'instant_bookable'---> lets create  t to 1 and f to 0
ab['instant_bookable']=np.where(ab['instant_bookable']=='t',1,0)
ab['instant_bookable'].value_counts()

# 'last_review'--> Since there are almost 21.4% values are Missing.
# replace missing with min
ab['last_review']=pd.to_datetime(ab['last_review'],format='%m/%d/%Y',errors='coerce').dt.year
ab.loc[ab['last_review'].isnull(),'last_review']=ab.loc[ab['data']=='train','last_review'].min()
ab['last_review'].value_counts()

#  'latitude' --> convert to x,y,z co-ordinates
#  'longitude' ---> convert to x,y,z co-ordinates
# After converting drop lat and long. Make x, y, z
x=np.cos(ab['latitude'])*np.cos(ab['longitude'])
y=np.cos(ab['latitude'])*np.sin(ab['longitude'])
z=np.sin(ab['latitude'])
lat_long=pd.DataFrame({'x':x,'y':y,'z':z})
lat_long.head()

ab.drop(['latitude','longitude'],axis=1,inplace=True)
ab=pd.concat([ab,lat_long],axis=1)

# 'review_scores_rating' --> Since there are almost 22.6% values are Missing.
# Replace  missing-->0
ab['review_scores_rating']=pd.to_numeric(np.where(ab['review_scores_rating'].isnull(),0,ab['review_scores_rating']))
ab['review_scores_rating'].value_counts()

# 'bedrooms'---> Replace missing value with median
ab.loc[ab['bedrooms'].isnull(),'bedrooms']=ab.loc[ab['data']=='train','bedrooms'].median()

# 'beds'---> Replace missing value with median
ab.loc[ab['beds'].isnull(),'beds']=ab.loc[ab['data']=='train','beds'].median()

cat_cols=ab.select_dtypes('object').columns
cat_cols

cat_cols=ab.select_dtypes('object').columns

for col in cat_cols[:-1]:
    freqs=ab[col].value_counts()
    k=freqs.index[freqs>20][:-1]
    for cat in k:
        name=col+'_'+cat
        ab[name]=(ab[col]==cat).astype(int)
    del ab[col]
    print(col)
        

ab.info()

ab.head()

# Data preprocessing complete - the data is in the expected format

ab.info()

ab.isnull().sum()

Let's separate our two data sets and remove the unnecessary columns that we added while combining them.

* Separate the train and test data
* We will break the train data further into train and validation datasets

#Separate the train and test data
ab_train=ab[ab['data']=='train']
del ab_train['data']

ab_test=ab[ab['data']=='test']
ab_test.drop(['log_price','data'],axis=1,inplace=True)

print('shape of train data after feature engineeing: {}'.format(ab_train.shape))
print('shape of test data after feature engineeing: {}'.format(ab_test.shape))

## Feature Scaling

feature_scale=[feature for feature in ab_train.columns if feature not in ['id','log_price']]

from sklearn.preprocessing import MinMaxScaler

scaler=MinMaxScaler()

scaler.fit(ab_train[feature_scale])

scaler.transform(ab_train[feature_scale])

# transform the train and add on the Id and SalePrice variables
scaled_features=pd.DataFrame(scaler.transform(ab_train[feature_scale]), columns=feature_scale)

data_train=pd.concat([ab_train[['id','log_price']].reset_index(drop=True),
                    scaled_features],axis=1)

data_train.to_csv('train.csv',index=False)

#### test data

feature_scale_test=[feature for feature in ab_test.columns if feature not in ['id']]
scaler.transform(ab_test[feature_scale_test])

# transform the test and add on the Id variables
scaled_features_test=pd.DataFrame(scaler.transform(ab_test[feature_scale_test]), columns=feature_scale_test)

data_test=pd.concat([ab_test[['id']].reset_index(drop=True),
                    scaled_features_test],axis=1)

data_test.to_excel('test.xlsx',index=False)
