==> AirBnB Price Prediction Challenge/2. Amenities Feature Engineering.ipynb <==
import pandas as pd
import numpy as np

# Read in data into dataframe
file_train=r'D:\Deloitte\case study\train.csv'
file_test=r'D:\Deloitte\case study\test.xlsx'

ab_train=pd.read_csv(file_train)
ab_test=pd.read_excel(file_test)

ab_train.head()

# text_data=['amenities']

amenities_train=ab_train[['amenities']]
amenities_test=ab_test[['amenities']]

print('train amenities data shape: {}'.format(amenities_train.shape))
print('test amenities data shape: {}'.format(amenities_test.shape))

amenities_train.columns

amenities_train.head()

from sklearn.preprocessing import MultiLabelBinarizer
mlb=MultiLabelBinarizer()

#### train data

train_temp=amenities_train['amenities'].str.replace('{','')

train_temp=train_temp.str.replace('}','')
train_temp=train_temp.str.replace('"','')
train_temp=train_temp.str.strip()
train_temp=train_temp.apply(lambda x: [k.strip() for k in x.strip().split(',') if
k!=''])

train_temp.head()

mlb.fit(train_temp)

mlb_train_amenities=pd.DataFrame(mlb.transform(train_temp), columns=mlb.classes_)
mlb_train_amenities

#### test data

test_temp=amenities_test['amenities'].str.replace('{','')

test_temp=test_temp.str.replace('}','')
test_temp=test_temp.str.replace('"','')
test_temp=test_temp.str.strip()
test_temp=test_temp.apply(lambda x: [k.strip() for k in x.strip().split(',') if
k!=''])

test_temp

mlb_test_amenities=pd.DataFrame(mlb.transform(test_temp), columns=mlb.classes_)
mlb_test_amenities

# # saving the 'aminities' train and test data
# mlb_train_amenities.to_csv('amenities_train.csv',index=False)
# mlb_test_amenities.to_excel('amenities_test.xlsx',index=False)

### Feature Scaling

from sklearn.preprocessing import MinMaxScaler
scaler=MinMaxScaler()

#### Amenities Train data

feature_scale_train=[feature for feature in mlb_train_amenities.columns]
scaler.fit(mlb_train_amenities[feature_scale_train])

# transform the train and add on the Id and SalePrice variables
scaled_aminities_train_features=pd.DataFrame(scaler.transform(mlb_train_amenities[feature_scale_train]), 
                                       columns=feature_scale_train)
scaled_aminities_train_features.head()

#### Amenities Test data

feature_scale_test=[feature for feature in mlb_test_amenities.columns]

# transform the test and add on the Id variables
scaled_aminities_test_features=pd.DataFrame(scaler.transform(mlb_test_amenities[feature_scale_test]),
                                                 columns=feature_scale_test)

scaled_aminities_test_features

# saving the 'aminities' train and test data
scaled_aminities_train_features.to_csv('amenities_train.csv',index=False)
scaled_aminities_test_features.to_excel('amenities_test.xlsx',index=False)
