==> AirBnB Price Prediction Challenge/4. Final Model_D.ipynb <==
# pandas and numpy for data manipulation
import numpy as np
import pandas as pd

# matplotlib and seaborn for visuilization
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

# Display all the columns of the dataframe
pd.pandas.set_option('display.max_columns',None)

# No warnings
import warnings
warnings.filterwarnings('ignore')

# Read  data into dataframe
file_train=r'D:\Deloitte\train.csv'
file_train_amenities=r'D:\Deloitte\amenities_train.csv'
file_test=r'D:\Deloitte\test.xlsx'
file_test_amenities=r'D:\Deloitte\amenities_test.xlsx'

ab_train=pd.read_csv(file_train)
amenities_train=pd.read_csv(file_train_amenities)
ab_test=pd.read_excel(file_test)
amenities_test=pd.read_excel(file_test_amenities)

# Returns the number of Rows and Columns in train data
print('shape of train data after preprocessing: {}'.format(ab_train.shape))
print('shape of amenities train data after preprocessing: {}'.format(amenities_train.shape))
print('shape of test data after preprocessing: {}'.format(ab_test.shape))
print('shape of amenities test data after preprocessing: {}'.format(amenities_test.shape))

# combining the amenities with other features of ab_train
train_ab=pd.concat([ab_train,amenities_train],axis=1)
print('shape of whole train data after concating: {}'.format(train_ab.shape))

test_ab=pd.concat([ab_test,amenities_test],axis=1)
print('shape of whole test data after concating: {}'.format(test_ab.shape))

#### Separating independent and dependent features

x_train=train_ab.drop(['id','log_price'],axis=1)
y_train=train_ab['log_price']

print('shape of y_train: {}'.format(x_train.shape))
print('shape of x_train: {}'.format(y_train.shape))

x_test=test_ab.drop(['id'],axis=1)

print('shape of x_test: {}'.format(x_test.shape))

* I choose to build XGBoost as my final Machine Learing Model

### 1: Machine Learning: XGBoost

from xgboost.sklearn import XGBRegressor
xgb_model=XGBRegressor()

xgb_model=XGBRegressor(**{'colsample_bytree': 0.6, 'gamma': 0.4, 'learning_rate': 0.01, 
                          'max_depth': 6, 'min_child_weight': 5, 'n_estimators': 1000, 
                          'reg_alpha': 5, 'scale_pos_weight': 3, 'subsample': 0.8})
xgb_model.fit(x_train,y_train)

test_predict=xgb_model.predict(x_test)

submissions=pd.DataFrame({'id':ab_test['id'],'log_price':test_predict})
submissions.to_excel('test__ML_price.xlsx',index=False)

submissions.head()

### 2: Deep learning: Neural Network Model

# Deep neural network

from keras.callbacks import ModelCheckpoint
from keras.models import Sequential
from keras.layers import Dense,Activation,Flatten,Dropout
import tensorflow as tf

NN_model=Sequential()

# Input layer:
NN_model.add(Dense(128,kernel_initializer='normal',input_dim=x_train.shape[1],activation='relu'))

# Hidden layers:
NN_model.add(Dense(256,kernel_initializer='normal',activation='relu'))
NN_model.add(Dropout(rate=0.2))
NN_model.add(Dense(256,kernel_initializer='normal',activation='relu'))
NN_model.add(Dropout(rate=0.2))
NN_model.add(Dense(256,kernel_initializer='normal',activation='relu'))

# output layer:
NN_model.add(Dense(1,kernel_initializer='normal',activation='relu'))

# Compile the network
NN_model.compile(loss='mean_squared_error',optimizer='adam',metrics=['mean_squared_error'])
NN_model.summary()

# checkpoint

checkpoint_name='weights-{epoch:03d}--{val_loss:.5f}.hdf5'
checkpoint=ModelCheckpoint(checkpoint_name,monitor='val_loss',verbose=1,save_best_only=True,
                          mode='auto')
callbacks_list=[checkpoint]

tf.config.run_functions_eagerly(True)

# Train the model
NN_model.fit(x_train,y_train.values,epochs=10,validation_split=0.2,callbacks=callbacks_list)

test_predict_NN=NN_model.predict(x_test)

submission=pd.DataFrame({'id':ab_test['id'],'log_price':test_predict_NN[:,0]})
submission.to_excel('test_NN_price.xlsx',index=False)

submission.head()
